Move hasSession & processDependents to shared
